Neat bot's training now can be controlled by changing dedicated variables in the constat.py file.
You have an ability to visualize the whole process of learning (Note that visualization is much slower that normal training)

You have an ability to set the speed of the training (Use it only for visualization, otherwise it has no sense)

You can decide whether you want to learn whole new generation or load the previous one that you have trained.

Note that your saved population will be erased when you start learning a new one.
 In case you want to copy the file which represents your population go to resources/population.dat